requirements.txt
Flask
pyaml
pytest